Export various snowflake driver interval funcs and types This is a squash of 9 commits:
==============================================================
Expose types and stash the raw response from snowflake.

We need this to be able to do fancy, snowflake specific things
with rows returned from db queries.
Note that this doesn't solve everything b/c sqlx.Rows is a struct
instead of an interface and doesn't expose the wrapped interface
directly, but that's a problem for another time.

Change-Id: I7b8619d6136ecd80500ad246e6844a89b5563332

==============================================================
Export setter methods for ExecResponse

Change-Id: Id1a47e262a1d1b73403a302f39e82366500a3c2d

==============================================================
Export SnowflakeRestful

Add a magic comment to use the specified request ID

Allows the caller to use that request ID to get updates on the query's
progress.

Change-Id: Ibd7086facdfd7cd4327e63593e3059ede3a59f02

==============================================================
Export SnowflakeRestful

With some reflection, we can use this type to post GET requests (and
other requests) to Snowflake ourselves.

Change-Id: Ia24d542b84edc948737f05ac421c706e1c2f9a7a

==============================================================
Fix bug in requestID parsing and simplify snowflakeRestful export

Fix some issues with the two previous commits:
1) remove the panic that I put in for testing (duh!). Luckily this panic
   does not break any existing clients.
2) use export.go to export snowflakeRestful, which is simpler and less
   intrusive

Change-Id: I3a30d41dd50aa15a490aa497b88e9635831524d9

==============================================================
4. Merge upstream changes

The main merge conflict resolved was to remove the hack of putting
REQUESTID in a comment in the SQL statement now that the upstream added
a way to specify the request ID in the context.

Change-Id: I6623ea1b23a648a4a259755ec073cde7f07dcaca

==============================================================
export types/method to support Arrow development

==============================================================
export simpleTokenAccessor for testing

snowflakeRestful.Token has been replaced by TokenAccessor. In testing,
we need to provide a fake TokenAccessor, which is simpleTokenAccessor.

Exporting it means we can reuse the fake in our unit tests.

==============================================================
exporting StringToValue function