Export SnowflakeRestful
This is a merge of the below 4 commits:

1. Add a magic comment to use the specified request ID

Allows the caller to use that request ID to get updates on the query's
progress.

Change-Id: Ibd7086facdfd7cd4327e63593e3059ede3a59f02

2. Export SnowflakeRestful

With some reflection, we can use this type to post GET requests (and
other requests) to Snowflake ourselves.

Change-Id: Ia24d542b84edc948737f05ac421c706e1c2f9a7a

3. Fix bug in requestID parsing and simplify snowflakeRestful export

Fix some issues with the two previous commits:
1) remove the panic that I put in for testing (duh!). Luckily this panic
   does not break any existing clients.
2) use export.go to export snowflakeRestful, which is simpler and less
   intrusive

Change-Id: I3a30d41dd50aa15a490aa497b88e9635831524d9

4. Merge upstream changes

The main merge conflict resolved was to remove the hack of putting
REQUESTID in a comment in the SQL statement now that the upstream added
a way to specify the request ID in the context.

Change-Id: I6623ea1b23a648a4a259755ec073cde7f07dcaca