Set default log level to warning
We can call SetLoggerLevel() in our code, but unfortunately there are a
few loggers in the init() functions in the gosnowflake library, so the
only way to avoid spam from there is to make a change to the library
itself.

Change-Id: I0af49fc636cbe827cf27982b8ce0ad83eeaf1b91

Downgrade logging of ignored errors

Some of the error log lines in ocsp.co literally say they are ignoring
the errors. So downgrade them from logger.Error to logger.Info to reduce
spam.

Change-Id: Ic519c2f7fdd93d2f4ad75d5735bf9483899c4cd2